Added Authentication and Authorization

+++ src/main/java/com/learningJavaBackend/projects/HotelBookingAndManagementSystem/Service/HotelServiceImpl.java
@@ -5,12 +5,15 @@ import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.DTO.Hote
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.DTO.RoomDto;
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Entity.Hotel;
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Entity.Room;
+import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Entity.User;
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Exception.ResourceNotFoundException;
+import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Exception.UnAuthorisedException;
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Repository.HotelRepository;
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -31,6 +34,10 @@ public class HotelServiceImpl implements HotelService{
         log.info("Creating new Hotel with name: {}", hotelDto.getName());
         Hotel hotel=modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+
+        User user= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel=hotelRepository.save(hotel);
         log.info("Created new Hotel with ID: {}", hotelDto.getId());
         return modelMapper.map(hotel, HotelDto.class);
@@ -42,6 +49,12 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel= hotelRepository
                 .findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID: " + id ));
+
+        User user= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with Id: " + id);
+        }
+
         return  modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -51,6 +64,12 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel= hotelRepository
                 .findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID: " + id ));
+
+        User user= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with Id: " + id);
+        }
+
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel=hotelRepository.save(hotel);
@@ -64,6 +83,11 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel= hotelRepository
                 .findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID: " + id ));
+
+        User user= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with Id: " + id);
+        }
 
         for(Room room: hotel.getRooms()){
             inventoryService.deleteAllInventories(room);
@@ -79,6 +103,11 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel= hotelRepository
                 .findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID: " + id ));
+
+        User user= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with Id: " + id);
+        }
 
         hotel.setActive(true);
 //        assuming only do it once
