+++ src/main/java/com/learningJavaBackend/projects/HotelBookingAndManagementSystem/Service/RoomServiceImpl.java
@@ -3,12 +3,15 @@ package com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Service
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.DTO.RoomDto;
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Entity.Hotel;
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Entity.Room;
+import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Entity.User;
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Exception.ResourceNotFoundException;
+import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Exception.UnAuthorisedException;
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Repository.HotelRepository;
 import com.learningJavaBackend.projects.HotelBookingAndManagementSystem.Repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -32,6 +35,11 @@ public class RoomServiceImpl implements RoomService{
                 .findById(hotelId)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID: " + hotelId ));
 
+        User user= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with Id: " + hotelId);
+        }
+
         Room room=modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room=roomRepository.save(room);
@@ -49,6 +57,11 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel= hotelRepository
                 .findById(hotelId)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID: " + hotelId ));
+
+        User user= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with Id: " + hotelId);
+        }
 
         return hotel.getRooms()
                 .stream()
@@ -73,6 +86,11 @@ public class RoomServiceImpl implements RoomService{
         Room room= roomRepository
                 .findById(roomId)
                 .orElseThrow(()->new ResourceNotFoundException("Room not found with ID: " + roomId ));
+
+        User user= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorisedException("This user does not own this room with Id: " + roomId);
+        }
 
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
